[scanner] fix: resolve lint baseline violation in AgentInstallGuide
Move mission file helpers out of AgentInstallGuide.tsx so the component file only exports React components, satisfying react-refresh/only-export-components.
